Day 19: Draw multiple PlayCursor/WriteCursor we store before
From what game draws when run, it looks like there are 2 PlayCursor
line for each 1 WriteCursor line. I guess that we write audio every
2 frames (one frame write data, one frame checks that we write too
much so it write nothing). It will be more favorable to write audio
for each frame to minimize audio delay, we should be "update the
past" on the the frame that we write nothing (?)